+++ CHANGELOG.md
@@ -17,6 +17,7 @@ List of bugs fixed:
 * 'Selection mode' keyboard shortcut did not work; not activate it with Shift + S (https://github.com/qupath/qupath/issues/638)
 * Exception when showing details for an extension that is missing a Manifest file (https://github.com/qupath/qupath/issues/664)
 * Exception when resetting an annotation description to an empty string (https://github.com/qupath/qupath/issues/661)
+* QuPath exits with a zero exit code if an exception is thrown from the command line (https://github.com/qupath/qupath/issues/654)
 
 ### Dependency updates*
 * Apache Commons Text 1.9